@academictelebotbyroshhellwett

+++ database/security_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, DateTime
+from sqlalchemy import Column, Integer, BigInteger, DateTime
 from database.db import Base  # Importing SHARED Base
 
 class UserStrike(Base):
@@ -8,7 +8,10 @@ class UserStrike(Base):
     """
     __tablename__ = "user_strikes"
     
-    user_id = Column(Integer, primary_key=True, index=True)
+    # CRITICAL FIX: Changed Integer -> BigInteger
+    # Telegram IDs (e.g., 7940390110) exceed the 32-bit Integer limit (2.1 billion)
+    user_id = Column(BigInteger, primary_key=True, index=True)
+    
     strike_count = Column(Integer, default=0)
     last_violation = Column(DateTime, nullable=True)
     #@academictelebotbyroshhellwett
